fix: one allele label everywhere, real regression test, clean lint and types
Tier 0: the groundwork that makes the earlier changes safe to build on.

One allele number, one label
----------------------------

`frontstr interpret` formatted allele numbers itself, from the older
allele_numeric/allele_numeric_source fields, while the report and exports used
the model's canonical number/number_method. The same allele therefore read as
"Δ-2" in the CLI and "14" in the report. Two numbers for one allele is not a
cosmetic problem in a forensic tool.

Allele.number_label now produces the label once, next to the number it
describes, and every view renders it. Allele.number_is_absolute moves the
"is this a real kit-comparable number" decision to the model as well.

Side effect: AMEL now prints X/Y instead of "len0". Its alleles carry their
designation in `isfg` and have no numeric allele, so the CLI's numeric fallback
was rendering a zero-length placeholder.

HG00113 in `frontstr interpret` before -> after:
  D3S1358  Δ-2      -> 14        vWA      Δ-3, Δ-1 -> 13, 17
  FGA      Δ2, Δ-1  -> 24, 21    D21S11   Δ2, 0    -> 35, 33
  AMEL     len0     -> X, Y

Regression test on real data
----------------------------

Every other test runs on synthetic fixtures, so the reference profile lived
only in session notes. tests/test_regression_hg00113.py runs the full pipeline
over the real ONT slice and asserts the genotype of all 25 markers, plus: no
dropouts, no false mixtures across all 5 single-source samples, POA active, and
TH01 9.3 keeping its interrupted structure.

The BAM slices are not versioned (~200 MB), so the tests skip cleanly when the
data is absent, which is the case in CI. Run locally with `-m integration`.
The two known nomenclature deviations (vWA, D21S11) are encoded as the expected
values with a note explaining why, so if a future change reconciles them the
test fails and says so rather than silently drifting.

CI green
--------

ruff 28 -> 0, mypy 14 -> 0. Mostly mechanical: bare dict/list generics, unused
imports, redundant int(round(...)), a re.search that can return None.

Ambiguous-unicode findings are configured away rather than fixed:
"±100 bp", "12 × 4" and "period −1" are the notation this domain uses, and
mangling them to ASCII makes the comments worse than the confusable characters
ever did.

+++ frontstr/report/ngs_display.py
@@ -62,10 +62,10 @@ def repeat_group_int(allele: dict[str, Any], period: int) -> int:
     """Repeat-count bin for chart X-axis and isoallele grouping."""
     an = allele.get("allele_numeric")
     if an is not None:
-        return int(round(float(an)))
+        return round(float(an))
     ce = allele.get("ce")
     if ce is not None:
-        return int(round(float(ce)))
+        return round(float(ce))
     isfg = str(allele.get("isfg") or "")
     match = _BRACKET_REPEAT.search(isfg)
     if match:
